refactor: split out ZagController
